requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-allauth==0.57.0
Pillow==10.1.0
qrcode==7.4.2
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
django-extensions==3.2.3
django-crispy-forms==2.1
crispy-bootstrap5==0.7
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
django-filter==23.5
django-import-export==3.3.1
django-storages==1.14.2
boto3==1.34.0
